add histogram intersection calculation for QiskitConnector

+++ org.planqk.nisq.analyzer.core/src/main/java/org/planqk/nisq/analyzer/core/model/ExecutionResult.java
@@ -70,7 +70,7 @@ public class ExecutionResult extends HasId {
 
     @Getter
     @Setter
-    private int histogramIntersectionValue;
+    private double histogramIntersectionValue;
 
     @Getter
     @Setter
