+++ org.planqk.nisq.analyzer.core/src/main/java/org/planqk/nisq/analyzer/core/web/dtos/entities/ExecutionResultDto.java
@@ -55,7 +55,7 @@ public class ExecutionResultDto extends RepresentationModel<ExecutionResultDto> 
 
     @Getter
     @Setter
-    private int histogramIntersectionValue;
+    private double histogramIntersectionValue;
 
     public static final class Converter {
